backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-ratelimit==4.1.0
django-redis==5.4.0
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
openai==1.40.6
chromadb==0.5.5
pypdf==4.3.1
gunicorn==22.0.0
httpx==0.27.0
whitenoise==6.6.0
